% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\name{create.experiment}
\alias{create.experiment}
\title{Create a structure experimental}
\usage{
create.experiment(
  ExpDesign = "exemple",
  wd = NULL,
  VOC = c("L", "B", "N"),
  control = "C",
  nS = 5,
  nC = 4
)
}
\arguments{
\item{ExpDesign}{Name of experiment}

\item{wd}{working directory}

\item{VOC}{VOC names}

\item{control}{control name}

\item{nS}{number of pre-drawn samples}

\item{nC}{number of pre-drawn concentration for each sample}
}
\value{
a csv file
}
\description{
Genere une structure correct pour l'utilisation du package. Pour les détails,
regardez create.empty.exp.design.
}
\examples{
create.experiment(VOC = c("L","B","N"), control = "C", nS = 5, nC = c(0.1,1,10,100))
}
