% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\docType{class}
\name{eag-class}
\alias{eag-class}
\title{object used for rEAG package}
\value{
a eag object
}
\description{
object used for rEAG package
}
\section{Slots}{

\describe{
\item{\code{depol}}{data.frame.}

\item{\code{eag}}{data.table.}

\item{\code{tmP}}{numeric.}

\item{\code{wd}}{character.}

\item{\code{names}}{character.}
}}

\examples{
#class(eag)
}
