% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\name{eag.compilation}
\alias{eag.compilation}
\title{Auto merge EAG}
\usage{
eag.compilation(
  expdes,
  control = "T",
  tmP = 50,
  tmD = NULL,
  ws = 25,
  wd = NULL,
  print.graph.total = TRUE
)
}
\arguments{
\item{expdes}{name of experiment in the folder 'experimental_design'}

\item{control}{name of control (concentration zero)}

\item{tmP}{duration of pulse (in centi second)}

\item{tmD}{duration of depolarisation (in centi second)}

\item{ws}{width selection}

\item{wd}{working directory}

\item{print.graph.total}{logical. Print or not a graph for view depolarisation, group by seq and by concentration}
}
\value{
a S4 eag object
}
\description{
Auto merge EAG
}
\examples{
# In 'experimental_design' folder, there is a ExpTestOO.csv corresponding at the experimental_design.
# Another folder, which named 'ExpTestOO' contains datas.
# ExpTestOO <- eag.compilation(expdes = "ExpTestOO", tmD = 200)
}
