% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\name{eag.import}
\alias{eag.import}
\title{Import EAG}
\usage{
eag.import(Sname = filenames[1], control = "T", tmP = 50, wd = NULL, ws = 25)
}
\arguments{
\item{Sname}{name of file.csv}

\item{control}{name of control (concentration zero)}

\item{tmP}{duration of pulse (in centi second)}

\item{wd}{working directory}

\item{ws}{width selection}
}
\value{
a S4 eag object
}
\description{
Importe un fichier csv, detecte les pulses de COV puis centre chaque pulse sur un T0.
On calcule l'intensite et le temps de depolarisation pour chaque pulse.
}
\examples{
# B_137 <- eag.import("Bombus137_C_0ppb_1h", control = "T", tmP = 50, wd = NULL)
}
