% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\name{eag.print}
\alias{eag.print}
\title{Print EAG}
\usage{
eag.print(eag, moda = "seq")
}
\arguments{
\item{eag}{a eag objet}

\item{moda}{a modality for group samples}
}
\value{
a figure
}
\description{
a simple function for print results of EAG
}
\examples{
# Meagdepol$con_seq <- paste0(Meag@depol$seq,"_",Meag@depol$con)
# eag.print(Meag, "con_seq")
}
