% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rEAG.R
\name{import.exp.design}
\alias{import.exp.design}
\title{Import experimental design}
\usage{
import.exp.design(ExpDes = "Experimental_Design_Name", wd = NULL)
}
\arguments{
\item{ExpDes}{file name (without extension)}

\item{wd}{working directory}
}
\value{
a data frame
}
\description{
Import un plan d'experience, les concentrations des VOC, les sequences,
modalites et commentaires
}
\examples{
# expdes <- import.exp.design("Experimental_design", wd = wod)
}
